// types.ts
export type NewsCategory = 'National' | 'International' | 'Sports' | 'Politics' | 'Economy' | 'Technology' | 'Health';

export interface NewsItem {
  id: string;
  title: string;
  summary: string;
  category: NewsCategory;
  imageUrl: string;
  date: string;
  timestamp: number;
  url?: string;
}